services/tts-worker/requirements.txt
# TTS
kokoro-onnx==0.1.0
onnxruntime==1.17.0
soundfile==0.12.1
numpy==1.26.4

# NATS
nats-py==2.7.0

# S3 Storage
boto3==1.34.0

# Data handling
pydantic==2.6.0
pydantic-settings==2.2.0

# Observability
prometheus-client==0.20.0
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-exporter-otlp==1.23.0
structlog==24.1.0

# Audio processing
pydub==0.25.1
ffmpeg-python==0.2.0

# Utilities
python-dotenv==1.0.1
tenacity==8.2.3

# Testing
pytest==8.0.0
pytest-cov==4.1.0
pytest-asyncio==0.23.0
